Move database functions to `storage`.
