+++ lib/main.js
@@ -2,4 +2,5 @@
  * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
  */
 
-export {OAuth2Client, checkDatabase, insertToken} from './oauth2-client.js';
+export {OAuth2Client} from './oauth2-client';
+export {get, insert} from './storage';
