+++ lib/oauth2-client.js
@@ -1,32 +1,11 @@
 /*!
  * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
  */
-import database from 'bedrock-mongodb';
+import * as storage from './storage';
 import * as bedrock from 'bedrock';
 import {httpClient} from '@digitalbazaar/http-client';
-import logger from './logger.js';
+import logger from './logger';
 const {util: {BedrockError}} = bedrock;
-
-bedrock.events.on('bedrock-mongodb.ready', async () => {
-  await database.openCollections(['access-tokens']);
-
-  await database.createIndexes([{
-    // prevent duplicate tokens from being stored
-    collection: 'access-tokens',
-    fields: {'token.token': 1},
-    options: {unique: true, background: false}
-  }, {
-    // automatically expire tokens with an `expires` date field
-    collection: 'access-tokens',
-    fields: {'token.expires': 1},
-    options: {
-      partialFilterExpression: {'token.expires': {$exists: true}},
-      unique: false,
-      background: false,
-      expireAfterSeconds: 0
-    }
-  }]);
-});
 
 export class OAuth2Client {
 /**
@@ -62,67 +41,16 @@ export class OAuth2Client {
       maxRetries
     } = this;
     let token;
-    token = await checkDatabase({client_id});
+    token = await storage.get({client_id});
     if(!token || !token.access_token) {
       token = await getNewAccessToken({
         client_id, client_secret, token_endpoint, grant_type, scope, httpsAgent,
         maxRetries
       });
-      await insertToken({token});
+      await storage.insert({token});
     }
     return token.access_token;
   }
-}
-
-/**
- * Inserts a token into the database.
- *
- * @param {object} options - Options to use.
- * @param {object} options.token - The token record, including client id and
- *    access token.
- *
- * @returns {string} The token record.
- */
-export async function insertToken({token}) {
-  const now = Date.now();
-  const meta = {created: now, updated: now};
-  // expires after 2 weeks
-  const ttl = 2 * 7 * 24 * 60 * 60 * 1000;
-  token.expires = new Date(now + ttl);
-  let record = {
-    meta,
-    token
-  };
-  const collection = database.collections['access-tokens'];
-  try {
-    const result = await collection.insertOne(record, database.writeOptions);
-    record = result.ops[0];
-  } catch(e) {
-    if(!database.isDuplicateError(e)) {
-      throw e;
-    }
-  }
-  return record;
-}
-
-/**
- * Queries the database for an existing access token for the client id.
- *
- * @param {object} options - Options to use.
- * @param {string} options.client_id - The ID of the client.
- *
- * @returns {string} The access token if there is one, null if not.
- */
-export async function checkDatabase({client_id}) {
-  const query = {
-    'token.client_id': client_id
-  };
-  let token;
-  const result = await database.collections['access-tokens'].findOne(query);
-  if(result && result.token) {
-    token = result.token;
-  }
-  return token;
 }
 
 /**
